// ClientApp_old/src/services/openapi/api.ts
export interface AddRecipeDiaryRequest {
    /**
     * 
     * @type {string}
     * @memberof AddRecipeDiaryRequest
     */
    'recipeId': string;
    /**
     * 
     * @type {MealTypes}
     * @memberof AddRecipeDiaryRequest
     */
    'mealType': MealTypes;
    /**
     * 
     * @type {number}
     * @memberof AddRecipeDiaryRequest
     */
    'quantity': number;
    /**
     * 
     * @type {string}
     * @memberof AddRecipeDiaryRequest
     */
    'entryDate': string;
}


/**
 * 
 * @export
 * @interface CreateRecipeRequest
 */
export interface CreateRecipeRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateRecipeRequest
     */
    'name': string;
}
/**
 * 
 * @export
 * @enum {string}
 */

export const DiaryType = {
    FoodDiary: 'FoodDiary',
    RecipeDiary: 'RecipeDiary'
} as const;

export type DiaryType = typeof DiaryType[keyof typeof DiaryType];


/**
 * 
 * @export
 * @interface Energy
 */
export interface Energy {
    /**
     * 
     * @type {string}
     * @memberof Energy
     */
    'unit': string;
    /**
     * 
     * @type {number}
     * @memberof Energy
     */
    'value': number;
}
// ...
export interface FoodDto {
    /**
     * 
     * @type {string}
     * @memberof FoodDto
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof FoodDto
     */
    'type': string;
    /**
     * 
     * @type {string}
     * @memberof FoodDto
     */
    'brandName': string;
    /**
     * 
     * @type {string}
     * @memberof FoodDto
     */
    'countryCode': string;
    /**
     * 
     * @type {string}
     * @memberof FoodDto
     */
    'name': string;
    /**
     * 
     * @type {NutritionalContent}
     * @memberof FoodDto
     */
    'nutritionalContents': NutritionalContent;
    /**
     * 
     * @type {{ [key: string]: ServingSizeDto; }}
     * @memberof FoodDto
     */
    'servingSizes': { [key: string]: ServingSizeDto; };
}
// ...
export interface IngredientDto {
    /**
     * 
     * @type {string}
     * @memberof IngredientDto
     */
    'id': string;
    /**
     * 
     * @type {FoodDto}
     * @memberof IngredientDto
     */
    'food': FoodDto;
    /**
     * 
     * @type {ServingSizeDto}
     * @memberof IngredientDto
     */
    'servingSize': ServingSizeDto;
    /**
     * 
     * @type {number}
     * @memberof IngredientDto
     */
    'quantity': number;
}
/**
 * 
 * @export
 * @interface LogInUserRequest
 */
export interface LogInUserRequest {
    /**
     * 
     * @type {string}
     * @memberof LogInUserRequest
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof LogInUserRequest
     */
    'password': string;
}
/**
 * 
 * @export
 * @enum {string}
 */

export const MealTypes = {
    Breakfast: 'Breakfast',
    Lunch: 'Lunch',
    Dinner: 'Dinner',
    Snacks: 'Snacks'
} as const;

export type MealTypes = typeof MealTypes[keyof typeof MealTypes];
// ...
export interface NutritionalContent {
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'calcium': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'carbohydrates': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'cholesterol': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'fat': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'fiber': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'iron': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'monounsaturatedFat': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'netCarbs': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'polyunsaturatedFat': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'potassium': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'protein': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'saturatedFat': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'sodium': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'sugar': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'transFat': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'vitaminA': number;
    /**
     * 
     * @type {number}
     * @memberof NutritionalContent
     */
    'vitaminC': number;
    /**
     * 
     * @type {Energy}
     * @memberof NutritionalContent
     */
    'energy': Energy;
}
// ...
export interface RecipeDto {
    /**
     * 
     * @type {string}
     * @memberof RecipeDto
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof RecipeDto
     */
    'name': string;
    /**
     * 
     * @type {Array<IngredientDto>}
     * @memberof RecipeDto
     */
    'ingredients': Array<IngredientDto>;
    /**
     * 
     * @type {NutritionalContent}
     * @memberof RecipeDto
     */
    'nutritionalContents': NutritionalContent;
}
// ...
export interface ServingSizeDto {
    /**
     * 
     * @type {string}
     * @memberof ServingSizeDto
     */
    'id': string;
    /**
     * 
     * @type {number}
     * @memberof ServingSizeDto
     */
    'nutritionMultiplier': number;
    /**
     * 
     * @type {number}
     * @memberof ServingSizeDto
     */
    'value': number;
    /**
     * 
     * @type {string}
     * @memberof ServingSizeDto
     */
    'unit': string;
}
// ...
export interface UpdateRecipeDiaryRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdateRecipeDiaryRequest
     */
    'id': string;
    /**
     * 
     * @type {number}
     * @memberof UpdateRecipeDiaryRequest
     */
    'quantity': number;
    /**
     * 
     * @type {MealTypes}
     * @memberof UpdateRecipeDiaryRequest
     */
    'mealType': MealTypes;
}
// ...
export class RecipeDiariesApi extends BaseAPI {
    /**
     * 
     * @param {AddRecipeDiaryRequest} addRecipeDiaryRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof RecipeDiariesApi
     */
    public addRecipeDiary(addRecipeDiaryRequest: AddRecipeDiaryRequest, options?: RawAxiosRequestConfig) {
        return RecipeDiariesApiFp(this.configuration).addRecipeDiary(addRecipeDiaryRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string | null} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof RecipeDiariesApi
     */
    public deleteRecipeDiary(id: string | null, options?: RawAxiosRequestConfig) {
        return RecipeDiariesApiFp(this.configuration).deleteRecipeDiary(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string | null} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof RecipeDiariesApi
     */
    public getRecipeDiaryById(id: string | null, options?: RawAxiosRequestConfig) {
        return RecipeDiariesApiFp(this.configuration).getRecipeDiaryById(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {UpdateRecipeDiaryRequest} updateRecipeDiaryRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof RecipeDiariesApi
     */
    public updateRecipeDiary(updateRecipeDiaryRequest: UpdateRecipeDiaryRequest, options?: RawAxiosRequestConfig) {
        return RecipeDiariesApiFp(this.configuration).updateRecipeDiary(updateRecipeDiaryRequest, options).then((request) => request(this.axios, this.basePath));
    }
}